+++ Simbody/src/SimbodyMatterSubsystem.cpp
@@ -1927,10 +1927,14 @@ int SimbodyMatterSubsystem::getTotalQAlloc()    const {return getRep().getTotalQ
 // Modeling info.
 void SimbodyMatterSubsystem::setUseEulerAngles(State& s, bool useAngles) const
   { getRep().setUseEulerAngles(s,useAngles); }
+void SimbodyMatterSubsystem::setUseEulerAnglesByDefault(bool useAngles)
+  { updRep().setUseEulerAnglesByDefault(useAngles); }
 void SimbodyMatterSubsystem::setConstraintIsDisabled(State& s, ConstraintIndex constraint, bool disabled) const
   { getRep().setConstraintIsDisabled(s,constraint,disabled); }
 bool SimbodyMatterSubsystem::getUseEulerAngles(const State& s) const
   { return getRep().getUseEulerAngles(s); }
+bool SimbodyMatterSubsystem::getUseEulerAnglesByDefault() const
+  { return getRep().getUseEulerAnglesByDefault(); }
 bool SimbodyMatterSubsystem::isConstraintDisabled(const State& s, ConstraintIndex constraint) const
   { return getRep().isConstraintDisabled(s,constraint); }
 void SimbodyMatterSubsystem::convertToEulerAngles(const State& inputState, State& outputState) const