+++ Simbody/tests/TestAngleConversions.cpp
@@ -56,9 +56,7 @@ int main() {
     lastBody = MobilizedBody::Weld(lastBody, Transform(Vec3(0, 0, 0)), body, Transform(Vec3(random.getValue(), random.getValue(), random.getValue())));
     lastBody = MobilizedBody::Screw(lastBody, Transform(Vec3(0, 0, 0)), body, Transform(Vec3(random.getValue(), random.getValue(), random.getValue())), 0.5);
     lastBody = MobilizedBody::Ellipsoid(lastBody, Transform(Vec3(0, 0, 0)), body, Transform(Vec3(random.getValue(), random.getValue(), random.getValue())));
-    mbs.realizeTopology();
-    State& s = mbs.updDefaultState();
-    mbs.realizeModel(s);
+    State s = mbs.realizeTopology();
     
     // Choose a random initial conformation.